+++ src/bract/core/inducer.clj
@@ -12,7 +12,35 @@
   (:require
     [bract.core.config :as config]
     [bract.core.echo   :as echo]
-    [bract.core.util   :as util]))
+    [bract.core.util   :as util])
+  (:import
+    [bract.core Echo]))
+
+
+(defn set-verbosity
+  "Set Bract verbosity flag and return context."
+  [context]
+  (Echo/setVerbose (config/ctx-verbose? context))
+  context)
+
+
+(defn read-config
+  "Use config filenames in the context to read and resolve config, and populate the context with it."
+  [context]
+  (let [config-files (config/ctx-config-files context)]
+    (if (seq config-files)
+      (->> config-files
+        (config/resolve-config context)
+        (assoc context (key config/ctx-config)))
+      context)))
+
+
+(defn run-inducers
+  "Run the inducers specified in the application config."
+  [context]
+  (->> (config/ctx-config context)
+    config/cfg-inducers
+    (util/induce context config/apply-inducer-by-name)))
 
 
 (defn context-hook