+++ src/bract/core/config.clj
@@ -23,7 +23,7 @@
 
 (keypin/defkey  ; context keys
   ctx-verbose?      [:bract.core/verbose? kputil/bool? "Verbose initialization?" {:parser  kputil/any->bool
-                                                                                  :default true
+                                                                                  :default false
                                                                                   :envvar  "APP_VERBOSE"
                                                                                   :sysprop "app.verbose"}]
   ctx-config-files  [:bract.core/config-files  vector? "Config file names"       {:parser  kputil/any->vec